product-ranking/requirements.txt
boto==2.38.0
cffi==1.1.2
colorama==0.3.3
cryptography==0.9.1
cssselect==0.9.1
enum34==1.0.4
Flask==0.10.1
Ghost.py==0.1.2
hjson==1.5.0
idna==2.0
ipaddress==1.0.7
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.4
MarkupSafe==0.23
mmh3==2.3.1
ndg-httpsclient==0.4.0
numpy==1.9.2
Pillow==2.8.2
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
pytesseract==0.1.6
PyYAML==3.11
queuelib==1.2.2
redis==2.10.3
requests==2.7.0
Scrapy==0.24.2
simmetrica==1.0.2
six==1.9.0
tldextract==1.6
Twisted==15.2.1
w3lib==1.11.0
Werkzeug==0.10.4
wheel==0.24.0
zope.interface==4.1.2
python-dateutil
psutil